% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_incidence.R
\name{get_prev}
\alias{get_prev}
\title{get_cohort}
\usage{
get_prev(fu_startd, data = shift_data, diseases = NULL, removeNA = FALSE)
}
\arguments{
\item{fu_startd}{data frame with exactly two columns: 'ergoid' and 'fu_startd'. ergoid  contains
the ids of participants for which the incidence/prevalence should be re-defined. It can also be a
subset of participants from RS-I through RS-III. fu_startd contains the baseline dates
corresponding to each participant. Participant with baseline dates set to NA will be automatically
assigned NA to all diseases.

ergoid should be formatted as factor, fu_startd needs to be a date in the yyyy-mm-dd format}

\item{data}{ERGO multimorbidity data. The default is shift_data.}

\item{diseases}{a character vector with diseases for which the baseline prevalence should be re-defined. By
default all diseases included in the multimorbidity data will be considered. A subset of the
diseases must be defined using a the following abbreviations:

dia: diabetes
stroke: stroke
can1: cancer
dem: dementia
hf: heart failure
chd: coronary heart disease
park: parkinsonism
dep1: depression
COPD: chronic obstructive pulmonary disease
asthma: asthma
tia: transient ischemic attack}

\item{removeNA}{whether participants with incomplete follow-up should be removed. The default is
set to FALSE, meaning data for all participants specified in fu_start_df will be returned.}
}
\description{
Takes the ERGO multimorbidity data and re-defines prevalence of diseases (so far)
included in the multimorbidity data. The prevalence is defined with respect to user defined
baseline date.
}
\details{
The output is cleaned data with ergoids, and prevalence indicator for each specified disease
(1 = prevalent case, 0 = disease free). As a bonus, the number of prevalent diseases at baseline
for each participant is also returned.

So far, the data include participants from  RS-I, RS-II and RS-III cohorts and the following
diseases:
\enumerate{
\item cancer (all types expect non-melanoma skin cancer)
\item COPD
\item asthma
\item stroke
\item diabetes
\item coronary heart disease
\item heart failure
\item TIA
\item dementia
\item parkinsonism
\item depression
}

To access the ERGO multimorbidity data, contact Frank van Rooij (f.vanrooij at erasmusmc.nl) and
Silvan Licher (s.licher at erasmusmc.nl)
}
